Delete World-Edit modpack.
This is required because W-E's license (AGPL) is highly viral,
even WORSE than the normal GPL. And compliance is very difficult.
Also, AGPL software cannot be bundled or in any way connected
with another, larger project without incurring risk of the whole
being termed 'derivative'. For this reason, AGLP software should
not be used, IMO.

The one function that the game made use of has been rewritten and
placed in the utility mod.