Delete chatplus mod.
This is required because the mod uses the GPLv3 license, which is
incompatible with the project's permissive licensing goals.

In any case, the mod has already been WHOLLY rewritten, so I may
simply remove the license and rename the mod, as long as I do it
prior to the public release of this game (as of this commit, the
game is not publicly released).