Prevent chest labels from being erased by accident
Since the `key_enter_field` is now also checked, and given that the
focus is right into the 'Label' field when a chest formspec is opened,
this avoids that a chest label get erased if a player accidentally hit
their 'Enter' key, by pre-filling the field with the current label.

Also helps when editing long chest labels, I guess.